Rename utils mod to fixture mod
